Moved PLAYLIST_AUTO_TYPE to __init__.py
