+++ hgtv/models/__init__.py
@@ -8,6 +8,29 @@ from hgtv import app
 db = SQLAlchemy(app)
 commentease = Commentease(db=db)
 
+class PLAYLIST_AUTO_TYPE:
+    WATCHED = 1
+    STARRED = 2
+    LIKED = 3
+    DISLIKED = 4
+    SPEAKING_IN = 5
+    APPEARING_IN = 6
+    CREW_IN = 7
+    ATTENDED = 8
+    QUEUE = 9
+
+playlist_auto_types = {
+    1: u"Watched",
+    2: u"Starred",
+    3: u"Liked",
+    4: u"Disliked",
+    5: u"Speaking in",
+    6: u"Appearing in",
+    7: u"Crew in",
+    8: u"Attended",
+    9: u"Queue",
+}
+
 from hgtv.models.video import *
 from hgtv.models.channel import *
 from hgtv.models.user import *
